backend/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
reportlab